Rebuild model to handle multiple moderation reviews
We need to handle a "2i moderator" and Business Approver so this splits
the person providing 2i moderation out into a separate model which
tracks how they moderated the Page.  Should we need more than one 2i
moderator in the future we should be able to handle that easily now we
have this split model.

The Business Approver field tracks an effectively-anonymous user.