GP2--3153-3154--new-opportunity-pages
This changeset is a squashed set of commits focused on adding two new page types to directory-cms, for the atlas project.

* A new InvestmentOpportunityPage (which replaces the concept of Capital Investment Opportunties and High Potential Opportunties being held in separate models, instead handling them both in the same page model)
* Its parent InvestmentOpportunityListPage.

These are similar to the detail and listing pages in the capital_invest module, but not identical. They do take much of the relationship modelling from them, so we can reuse as much code in great-international-ui as possible - we will try to repurpose much of the view and template code for capinvest for the atlas pages.

Other points to note:
* The new detail page uses a StreamField.
* Rich text is still done with markdown, which is rendered to HTML before it goes over the wire.
* Wagtail-modeltranslation has been configured for the new pages, and it appears to play nicely with streamfields. Note that for the MVP atlas, we won't have translations being used.
* The changeset also includes Serializers (and nested Serializers) for the new model, along with SOME tests, covering critical behaviour. Coverage was not at 100% before this work started, but we should aim to get it up later on.
* Migrations (10+) have been merged/squashed/recreated as a single migration, because running each migration in d-cms takes time
* Coverage.py is run by default with pytest now